+++ composable_kernel/include/gridwise_generic_reduction_wrapper_common.hpp
@@ -1,5 +1,5 @@
-#ifndef DYNAMIC_GRIDWISE_GENERIC_REDUCTION_WRAPPER_COMMON
-#define DYNAMIC_GRIDWISE_GENERIC_REDUCTION_WRAPPER_COMMON
+#ifndef CK_GRIDWISE_GENERIC_REDUCTION_WRAPPER_COMMON
+#define CK_GRIDWISE_GENERIC_REDUCTION_WRAPPER_COMMON
 
 #include "config.hpp"
 #include "number.hpp"
