+++ composable_kernel/include/utility/reduction_functions_binop.hpp
@@ -23,13 +23,13 @@
  * SOFTWARE.
  *
  *******************************************************************************/
-#ifndef CK_DYNAMIC_REDUCTION_FUNCTIONS_BINOP_HPP
-#define CK_DYNAMIC_REDUCTION_FUNCTIONS_BINOP_HPP
+#ifndef CK_REDUCTION_FUNCTIONS_BINOP_HPP
+#define CK_REDUCTION_FUNCTIONS_BINOP_HPP
 
 #include "data_type.hpp"
 
 #include "reduction_common.hpp"
-#include "dynamic_reduction_operator.hpp"
+#include "reduction_operator.hpp"
 
 namespace ck {
 namespace detail {
